move Newton parser to its own file
